refactor(BudgetDrawer): move the component to its own directory
