+++ src/components/parameters/budgets/AllBudgetsTable.tsx
@@ -9,7 +9,7 @@ import {
 } from '@mui/x-data-grid'
 
 import { BudgetEditType, BudgetResponseType } from '~types/budget'
-import BudgetDrawer from '~components/drawers/Transactions/BudgetDrawer'
+import BudgetDrawer from '~/components/drawers/Transactions/BudgetDrawer/BudgetDrawer'
 
 type AllBudgetsTableProps = {
     data: BudgetResponseType[] | undefined
