+++ src/pages/transacciones/presupuesto/index.tsx
@@ -11,7 +11,7 @@ import EditToolbar from '~components/table/headers/toolbar'
 import PageTitle from '~components/titles/PageTitle'
 import AllBudgetsTable from '~components/parameters/budgets/AllBudgetsTable'
 import { useGetAllBudgetsQuery } from '~redux/api/bca-backend/transacciones/budgetSlice'
-import BudgetDrawer from '~components/drawers/Transactions/BudgetDrawer'
+import BudgetDrawer from '~/components/drawers/Transactions/BudgetDrawer/BudgetDrawer'
 import { useGetAllProjectsQuery } from '~redux/api/bca-backend/parametros/projectsSlice'
 
 export default function Presupuesto() {
